requirements.txt
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
cssselect==1.2.0
cssutils==2.7.1
dnspython==2.4.2
email-validator==2.0.0.post2
flake8==6.1.0
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
lxml==4.9.3
mccabe==0.7.0
more-itertools==10.1.0
premailer==3.10.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
python-dotenv==1.0.0
requests==2.31.0
SecretStorage==3.3.3
urllib3==2.0.4
yagmail==0.15.293
zipp==3.16.2
